requirements.txt
pandas>=0.23.4
pyyaml>=3.13
sqlalchemy>=1.2.14
pymysql>=0.9.2